+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 - 0.0.0-beta1
-    Initial release
+    - Added `insert` method